+++ lib/src/exports/extensions/data-types/string_ext.dart
@@ -4,5 +4,7 @@ extension StringExt on String {
   bool isEquals(String val, {bool ignoreCase = false}) =>
       ignoreCase ? (toLowerCase() == val.toLowerCase()) : (this == val);
 
+  String get truellipses => replaceAll('', '\u200B');
+
   TextCreator get txt => TextCreator(data: this);
 }